+++ python/ray/tests/horovod/BUILD
@@ -1,0 +1,9 @@
+py_test(
+    name = "test_horovod",
+    size = "medium",
+    srcs = ["test_horovod.py"],
+    deps = [":horovod_lib"],
+    tags = ["team:ml", "exclusive"]
+)
+
+
