+++ python/ray/tests/lightgbm/BUILD
@@ -1,0 +1,15 @@
+py_test(
+ name = "simple_example",
+ size = "small",
+ srcs = ["simple_example.py"],
+ tags = ["team:ml", "exclusive"],
+)
+
+py_test(
+ name = "simple_tune",
+ size="small",
+ srcs = ["simple_tune.py"],
+ tags = ["team:ml", "exclusive"]
+)
+
+
