+++ python/ray/tests/lightgbm/simple_example.py
@@ -1,7 +1,7 @@
 from sklearn import datasets
 from sklearn.model_selection import train_test_split
 
-from ray.util.lightgbm import RayDMatrix, RayParams, train
+from lightgbm_ray import RayDMatrix, RayParams, train
 
 
 # __lightgbm_begin__
