+++ python/ray/util/horovod/__init__.py
@@ -1,0 +1,5 @@
+raise DeprecationWarning(
+    "ray.util.horovod has been removed as of Ray 2.0. Instead, use the `horovod` "
+    "library directly or the `HorovodTrainer` in Ray AIR ("
+    "https://docs.ray.io/en/master/ray-air/getting-started.html)"
+)
