+++ python/ray/util/lightgbm/__init__.py
@@ -1,37 +1,5 @@
-import logging
-
-logger = logging.getLogger(__name__)
-
-train = None
-predict = None
-RayParams = None
-RayDMatrix = None
-RayFileType = None
-RayLGBMClassifier = None
-RayLGBMRegressor = None
-
-try:
-    from lightgbm_ray import (
-        train,
-        predict,
-        RayParams,
-        RayDMatrix,
-        RayFileType,
-        RayLGBMClassifier,
-        RayLGBMRegressor,
-    )
-except ImportError:
-    logger.info(
-        "lightgbm_ray is not installed. Please run "
-        "`pip install git+https://github.com/ray-project/lightgbm_ray`."
-    )
-
-__all__ = [
-    "train",
-    "predict",
-    "RayParams",
-    "RayDMatrix",
-    "RayFileType",
-    "RayLGBMClassifier",
-    "RayLGBMRegressor",
-]
+raise DeprecationWarning(
+    "ray.util.lightgbm has been removed as of Ray 2.0. Instead, use the `lightgbm-ray` "
+    "library directly or the `LightGBMTrainer` in Ray AIR ("
+    "https://docs.ray.io/en/master/ray-air/getting-started.html)"
+)
