+++ release/ml_user_tests/horovod/horovod_user_test.py
@@ -3,7 +3,7 @@ import os
 import time
 
 import ray
-from ray.util.horovod.horovod_example import main
+from ray.tests.horovod.horovod_example import main
 
 if __name__ == "__main__":
     start = time.time()
